Hierarchical symbol table (#29)
* Make symbol table hierarchical. Works for tests not involving templates. Template handling code needs to be adjusted now.

* Move ir function name creating to CodegenContext because it's tied to the ir module.
 Makes test_scoped_function_redefinition work.

* Make new symbol table work with old template handling code

* Remove no longer needed used_ir_func_names

* Integrate current_scope_identifiers with the hierarchical symbol table so that is does not have to be managed separately.

Also, whether you start a new scope where variable identifiers of the parent are no longer visible is declarative and explicit now.

* Remove unused symbols member and identifier_in_scope method

* Move FunctionDeclarationAst into own file.

* Move ir building for concrete declared functions into subclass of ConcreteFunction.

This avoids implicit capturing of the symbol table and FunctionDeclarationAst. The integration with existing code is currently hacky and will improve once all ConcreteFunctionFactories are converted.

* Use subclass of function template to handle creation concrete functions.

Needed to fix copying to work correctly for subclasses as well. Hopefully won't need copies at all soon.

* Handle ctor and dtor without a function factory.

No subclasses of ConcreteFunction needed because ctors and dtors can not be inlined

* Implement builtin and bitcast functions in new way.

* Fix bitcast function