Make semicolon more optional (#38)
* Require ; only between statements without new lines.

New line is lexed and only ignored in the parser if it is not expected. New line can be used as a separator.

 Stmt and StmtList are adapted to only require separators between statements to allow e.g. func f() { return 5 }

 MemberList still needs to be adapted to allow struct S { i: Int }

The only statement that does not need a separator is the if statement (because we run into problems otherwise). Could/Should extend that to all statements ending with a Scope i.e. }

* DEATH TO THE SEMICOLONS (in the preamble)

... except the 15 of 165 that are actually necessary